requirements.txt
scapy
python-nmap
requests
tk
